First Day
The first day of computer science class 2020 started with a quick revision of contents from last year such as recursion, computer architecture, logic gate, and so on. Then, Dr. Ruben gave us a scavenger type of task which required us to go around the campus to figure out the answer using the knowledge we had about computer science. It was a fun activity to start off the year.